+++ android/expoview/src/main/java/host/exp/exponent/experience/HomeActivity.java
@@ -46,6 +46,7 @@ import host.exp.exponent.ExponentManifest;
 import host.exp.exponent.RNObject;
 import host.exp.exponent.analytics.Analytics;
 import host.exp.exponent.di.NativeModuleDepsProvider;
+import host.exp.exponent.experience.splashscreen.HomeSplashScreenViewProvider;
 import host.exp.exponent.kernel.ExperienceId;
 import host.exp.exponent.kernel.Kernel;
 import host.exp.exponent.utils.ExperienceActivityUtils;