+++ packages/expo-splash-screen/android/src/main/java/expo/modules/splashscreen/NativeResourcesBasedSplashScreenViewProvider.kt
@@ -1,7 +1,9 @@
 package expo.modules.splashscreen
 
 import android.content.Context
+import android.view.View
 import androidx.core.content.ContextCompat
+import expo.modules.splashscreen.singletons.SplashScreen
 
 // this needs to stay for versioning to work
 
@@ -12,7 +14,7 @@ class NativeResourcesBasedSplashScreenViewProvider(
   private val resizeMode: SplashScreenImageResizeMode
 ) : SplashScreenViewProvider {
 
-  override fun createSplashScreenView(context: Context): SplashScreenView {
+  override fun createSplashScreenView(context: Context): View {
     val splashScreenView = SplashScreenView(context)
     splashScreenView.setBackgroundColor(getBackgroundColor(context))