+++ packages/expo-splash-screen/android/src/main/java/expo/modules/splashscreen/SplashScreenListenerActivity.kt
@@ -1,0 +1,5 @@
+package expo.modules.splashscreen
+
+interface SplashScreenListenerActivity {
+    fun onSplashScreenDismissed(): Unit
+}