+++ packages/expo-splash-screen/ios/EXSplashScreen/EXSplashScreenService.m
@@ -2,6 +2,7 @@
 
 #import <EXSplashScreen/EXSplashScreenService.h>
 #import <EXSplashScreen/EXSplashScreenViewNativeProvider.h>
+#import <EXSplashScreen/EXSplashScreenWarningViewController.h>
 #import <UMCore/UMDefines.h>
 
 @interface EXSplashScreenService ()
@@ -31,6 +32,7 @@ UM_REGISTER_SINGLETON_MODULE(SplashScreen);
                    failureCallback:^(NSString *message){ UMLogWarn(@"%@", message); }];
 }
 
+
 - (void)showSplashScreenFor:(UIViewController *)viewController
    splashScreenViewProvider:(id<EXSplashScreenViewProvider>)splashScreenViewProvider
             successCallback:(void (^)(void))successCallback
@@ -42,6 +44,7 @@ UM_REGISTER_SINGLETON_MODULE(SplashScreen);
   
   EXSplashScreenController *splashScreenController = [[EXSplashScreenController alloc] initWithViewController:viewController
                                                                                      splashScreenViewProvider:splashScreenViewProvider];
+  
   [self.splashScreenControllers setObject:splashScreenController forKey:viewController];
   [[self.splashScreenControllers objectForKey:viewController] showWithCallback:successCallback
                                                                failureCallback:failureCallback];
@@ -66,6 +69,12 @@ UM_REGISTER_SINGLETON_MODULE(SplashScreen);
   if (![self.splashScreenControllers objectForKey:viewController]) {
     return failureCallback(@"No native splash screen registered for given view controller. Call 'SplashScreen.show' for given view controller first.");
   }
+  
+  if ([[viewController class] conformsToProtocol:@protocol(EXSplashScreenWarningViewController)]) {
+    UIViewController<EXSplashScreenWarningViewController> *warningVC = (UIViewController <EXSplashScreenWarningViewController> *)viewController;
+    [warningVC onSplashScreenDimissed];
+  }
+  
   return [[self.splashScreenControllers objectForKey:viewController] hideWithCallback:successCallback
                                                                       failureCallback:failureCallback];
 }