+++ packages/expo-splash-screen/ios/EXSplashScreen/EXSplashScreenWarningViewController.h
@@ -1,0 +1,19 @@
+//
+//  EXSplashScreenWarningViewController.h
+//  EXSplashScreen
+//
+//  Created by andrew on 2021-06-02.
+//
+
+#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@protocol EXSplashScreenWarningViewController
+
+- (void)onSplashScreenDimissed;
+
+@end
+
+NS_ASSUME_NONNULL_END